Final commit

+++ features/medical_checkup_details/handler/handler.go
@@ -10,73 +10,101 @@ import (
 )
 
 type MedicalCheckupDetailHandler struct {
-	s medicalcheckupdetails.MedicalCheckupDetailServiceInterface
+	s   medicalcheckupdetails.MedicalCheckupDetailServiceInterface
+	jwt helper.JWTInterface
 }
 
-func NewHandler(service medicalcheckupdetails.MedicalCheckupDetailServiceInterface) medicalcheckupdetails.MedicalCheckupDetailHandlerInterface {
+func NewHandler(service medicalcheckupdetails.MedicalCheckupDetailServiceInterface, jwt helper.JWTInterface) medicalcheckupdetails.MedicalCheckupDetailHandlerInterface {
 	return &MedicalCheckupDetailHandler{
-		s: service,
+		s:   service,
+		jwt: jwt,
 	}
 }
 
 func (mcdh *MedicalCheckupDetailHandler) GetMedicalCheckupDetails() echo.HandlerFunc {
 	return func(c echo.Context) error {
+		role := mcdh.jwt.CheckRole(c)
+
+		if role != "Admin" {
+			return c.JSON(http.StatusUnauthorized, helper.FormatResponse("Only admin can access this page", nil))
+		}
+
 		var paramMCUID = c.Param("idmcu")
 		idMcu, err := strconv.Atoi(paramMCUID)
 		if err != nil {
-			c.Logger().Fatal("Handler : Param ID MCU Error : ", err.Error())
-			return c.JSON(http.StatusBadRequest, helper.FormatResponse("Fail", nil))
+			return c.JSON(http.StatusBadRequest, helper.FormatResponse("Invalid param id medical checkup", nil))
 		}
 
-		result, err := mcdh.s.GetMedicalCheckupDetails(idMcu)
+		result, resMed, err := mcdh.s.GetMedicalCheckupDetails(idMcu)
 
 		if err != nil {
-			c.Logger().Fatal("Handler : Get All Process Error : ", err.Error())
-			return c.JSON(http.StatusInternalServerError, helper.FormatResponse("Fail", nil))
+			return c.JSON(http.StatusInternalServerError, helper.FormatResponse("Cannot process data", nil))
 		}
 
-		return c.JSON(http.StatusOK, helper.FormatResponse("Success", result))
+		response := MedicalCheckupResponseDetail{
+			MedicalCheckupID: result.MedicalCheckupID,
+			Complain:         result.Complain,
+			Treatment:        result.Treatment,
+			DetailInfo: make([]struct {
+				MedicineName string "json:\"medicine_name\""
+				Quantity     int    "json:\"quantity\""
+			}, len(resMed)),
+		}
+
+		for i, medicine := range resMed {
+			response.DetailInfo[i].MedicineName = medicine.MedicineName
+			response.DetailInfo[i].Quantity = medicine.Quantity
+		}
+
+		return c.JSON(http.StatusOK, helper.FormatResponse("Success get medical checkup detail", response))
 	}
 }
 func (mcdh *MedicalCheckupDetailHandler) GetMedicalCheckupDetail() echo.HandlerFunc {
 	return func(c echo.Context) error {
+		role := mcdh.jwt.CheckRole(c)
+
+		if role != "Admin" {
+			return c.JSON(http.StatusUnauthorized, helper.FormatResponse("Only admin can access this page", nil))
+		}
+
 		var paramMCUID = c.Param("idmcu")
 		idMcu, err := strconv.Atoi(paramMCUID)
 		if err != nil {
-			c.Logger().Fatal("Handler : Param ID MCU Error : ", err.Error())
-			return c.JSON(http.StatusBadRequest, helper.FormatResponse("Fail", nil))
+			return c.JSON(http.StatusBadRequest, helper.FormatResponse("Invalid param id medical checkup", nil))
 		}
 
 		var paramMCUDID = c.Param("idmcudetail")
 		idMcuDetail, err := strconv.Atoi(paramMCUDID)
 		if err != nil {
-			c.Logger().Fatal("Handler : Param ID MCU Detail Error : ", err.Error())
-			return c.JSON(http.StatusBadRequest, helper.FormatResponse("Fail", nil))
+			return c.JSON(http.StatusBadRequest, helper.FormatResponse("Invalid param id medical checkup detail", nil))
 		}
 
 		result, err := mcdh.s.GetMedicalCheckupDetail(idMcu, idMcuDetail)
 
 		if err != nil {
-			c.Logger().Fatal("Handler : Get By ID Process Error : ", err.Error())
-			return c.JSON(http.StatusInternalServerError, helper.FormatResponse("Fail", nil))
+			return c.JSON(http.StatusInternalServerError, helper.FormatResponse("Cannot process data", nil))
 		}
 
-		return c.JSON(http.StatusOK, helper.FormatResponse("Success", result))
+		return c.JSON(http.StatusOK, helper.FormatResponse("Success get medical checkup detail", result))
 	}
 }
 func (mcdh *MedicalCheckupDetailHandler) CreateMedicalCheckupDetail() echo.HandlerFunc {
 	return func(c echo.Context) error {
+		role := mcdh.jwt.CheckRole(c)
+
+		if role != "Admin" {
+			return c.JSON(http.StatusUnauthorized, helper.FormatResponse("Only admin can access this page", nil))
+		}
+
 		var paramMCUID = c.Param("idmcu")
 		idMcu, err := strconv.Atoi(paramMCUID)
 		if err != nil {
-			c.Logger().Fatal("Handler : Param ID MCU Error : ", err.Error())
-			return c.JSON(http.StatusBadRequest, helper.FormatResponse("Fail", nil))
+			return c.JSON(http.StatusBadRequest, helper.FormatResponse("Invalid param id medical checkup", nil))
 		}
 
 		var input = new(InputRequest)
 		if err := c.Bind(&input); err != nil {
-			c.Logger().Fatal("Handler : Bind Input Error  : ", err.Error())
-			return c.JSON(http.StatusBadRequest, helper.FormatResponse("Fail", nil))
+			return c.JSON(http.StatusBadRequest, helper.FormatResponse("Invalid user input", nil))
 		}
 
 		var serviceInput = new(medicalcheckupdetails.MedicalCheckupDetail)
@@ -87,8 +115,7 @@ func (mcdh *MedicalCheckupDetailHandler) CreateMedicalCheckupDetail() echo.Handl
 		result, err := mcdh.s.CreateMedicalCheckupDetail(*serviceInput)
 
 		if err != nil {
-			c.Logger().Fatal("Handler : Input Process Error : ", err.Error())
-			return c.JSON(http.StatusInternalServerError, helper.FormatResponse("Fail", nil))
+			return c.JSON(http.StatusInternalServerError, helper.FormatResponse("Cannot process data", nil))
 		}
 
 		var response = new(InputResponse)
@@ -96,29 +123,32 @@ func (mcdh *MedicalCheckupDetailHandler) CreateMedicalCheckupDetail() echo.Handl
 		response.MedicineID = result.MedicineID
 		response.Quantity = result.Quantity
 
-		return c.JSON(http.StatusCreated, helper.FormatResponse("Success", response))
+		return c.JSON(http.StatusCreated, helper.FormatResponse("Success created medical checkup detail", response))
 	}
 }
 func (mcdh *MedicalCheckupDetailHandler) UpdateMedicalCheckupDetail() echo.HandlerFunc {
 	return func(c echo.Context) error {
+		role := mcdh.jwt.CheckRole(c)
+
+		if role != "Admin" {
+			return c.JSON(http.StatusUnauthorized, helper.FormatResponse("Only admin can access this page", nil))
+		}
+
 		var paramMCUID = c.Param("idmcu")
 		idMcu, err := strconv.Atoi(paramMCUID)
 		if err != nil {
-			c.Logger().Fatal("Handler : Param ID MCU Error : ", err.Error())
-			return c.JSON(http.StatusBadRequest, helper.FormatResponse("Fail", nil))
+			return c.JSON(http.StatusBadRequest, helper.FormatResponse("Invalid param id medical checkup", nil))
 		}
 
 		var paramMCUDID = c.Param("idmcudetail")
 		idMcuDetail, err := strconv.Atoi(paramMCUDID)
 		if err != nil {
-			c.Logger().Fatal("Handler : Param ID MCU Detail Error : ", err.Error())
-			return c.JSON(http.StatusBadRequest, helper.FormatResponse("Fail", nil))
+			return c.JSON(http.StatusBadRequest, helper.FormatResponse("Invalid param id medical checkup detail", nil))
 		}
 
 		var input = new(InputRequest)
 		if err := c.Bind(&input); err != nil {
-			c.Logger().Fatal("Handler : Bind Input Error  : ", err.Error())
-			return c.JSON(http.StatusBadRequest, helper.FormatResponse("Fail", nil))
+			return c.JSON(http.StatusBadRequest, helper.FormatResponse("Invalid user input", nil))
 		}
 
 		var serviceInput = new(medicalcheckupdetails.UpdateMedicalCheckupDetail)
@@ -128,36 +158,38 @@ func (mcdh *MedicalCheckupDetailHandler) UpdateMedicalCheckupDetail() echo.Handl
 		result, err := mcdh.s.UpdateMedicalCheckupDetail(*serviceInput, idMcu, idMcuDetail)
 
 		if err != nil {
-			c.Logger().Fatal("Handler : Update Process Error : ", err.Error())
-			return c.JSON(http.StatusInternalServerError, helper.FormatResponse("Fail", nil))
+			return c.JSON(http.StatusInternalServerError, helper.FormatResponse("Cannot process data", nil))
 		}
 
-		return c.JSON(http.StatusOK, helper.FormatResponse("Success", result))
+		return c.JSON(http.StatusOK, helper.FormatResponse("Success update medical checkup detail", result))
 	}
 }
 func (mcdh *MedicalCheckupDetailHandler) DeleteMedicalCheckupDetail() echo.HandlerFunc {
 	return func(c echo.Context) error {
+		role := mcdh.jwt.CheckRole(c)
+
+		if role != "Admin" {
+			return c.JSON(http.StatusUnauthorized, helper.FormatResponse("Only admin can access this page", nil))
+		}
+
 		var paramMCUID = c.Param("idmcu")
 		idMcu, err := strconv.Atoi(paramMCUID)
 		if err != nil {
-			c.Logger().Fatal("Handler : Param ID MCU Error : ", err.Error())
-			return c.JSON(http.StatusBadRequest, helper.FormatResponse("Fail", nil))
+			return c.JSON(http.StatusBadRequest, helper.FormatResponse("Invalid param id medical checkup", nil))
 		}
 
 		var paramMCUDID = c.Param("idmcudetail")
 		idMcuDetail, err := strconv.Atoi(paramMCUDID)
 		if err != nil {
-			c.Logger().Fatal("Handler : Param ID MCU Detail Error : ", err.Error())
-			return c.JSON(http.StatusBadRequest, helper.FormatResponse("Fail", nil))
+			return c.JSON(http.StatusBadRequest, helper.FormatResponse("Invalid param id medical checkup detail", nil))
 		}
 
 		result, err := mcdh.s.DeleteMedicalCheckupDetail(idMcu, idMcuDetail)
 
 		if err != nil {
-			c.Logger().Fatal("Handler : Delete Process Error : ", err.Error())
-			return c.JSON(http.StatusInternalServerError, helper.FormatResponse("Fail", nil))
+			return c.JSON(http.StatusInternalServerError, helper.FormatResponse("Cannot process data", nil))
 		}
 
-		return c.JSON(http.StatusNoContent, helper.FormatResponse("Success", result))
+		return c.JSON(http.StatusNoContent, helper.FormatResponse("Success delete medical checkup detail", result))
 	}
 }
